Just remove bond configuration when they exist.
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=995997

+++ src/ovirt/node/config/defaults.py
@@ -299,7 +299,8 @@ class Network(NodeConfigFileSection):
                         NicConfig(bifname).delete()
 
                 bonds = Bonds()
-                bonds.delete_all()
+                if bonds.is_enabled():
+                    bonds.delete_all()
 
             def _remove_ifcfg_configs(self):
                 pat = NicConfig.IfcfgBackend.filename_tpl % "*"
